+++ src/components/Find/Find.scss
@@ -1,34 +1,37 @@
+/* stylelint-disable max-nesting-depth */
+/* stylelint-disable selector-class-pattern */
 .nav-center {
-  width: 300px;
   position: relative;
+  width: 300px;
+  margin: 0 10px;
 
   input {
-    outline: none;
-    border: #dbdbdb 1px solid;
-    color: #9d9d9d;
-    padding: 5px;
-    border-radius: 5px;
-    width: 100%;
     box-sizing: border-box;
+    width: 100%;
+    padding: 5px;
+    border: #dbdbdb 1px solid;
+    border-radius: 5px;
+    outline: none;
+    color: #9d9d9d;
     text-align: center;
   }
   .idBox {
-    background: #fff;
-    border-radius: 5px;
     position: absolute;
-    border: 1px solid #dbdbdb;
-    width: 100%;
-    height: 300px;
-    bottom: -300px;
-    box-sizing: border-box;
-    overflow: scroll;
     // display: none;
     z-index: 100;
+    bottom: -300px;
+    box-sizing: border-box;
+    width: 100%;
+    height: 300px;
+    overflow: scroll;
+    border: 1px solid #dbdbdb;
+    border-radius: 5px;
+    background: #fff;
 
     .id-top {
       display: flex;
-      justify-content: space-between;
       align-items: center;
+      justify-content: space-between;
 
       p {
         margin: 0;
@@ -36,32 +39,32 @@
       }
 
       .search {
-        padding-left: 10px;
         box-sizing: border-box;
-        font-size: 0.8rem;
         margin: 0;
+        padding-left: 10px;
+        font-size: 0.8rem;
       }
 
       .delete {
-        font-size: 0.6rem;
-        color: rgb(7, 96, 230);
-        cursor: pointer;
         padding-right: 10px;
+        color: rgb(7, 96, 230);
+        font-size: 0.6rem;
+        cursor: pointer;
       }
     }
 
     .idprofile {
-      list-style: none;
       padding-left: 0;
+      list-style: none;
 
       li {
         display: flex;
         align-items: center;
         justify-content: start;
-        border-bottom: 1px solid #9d9d9d;
         width: 100%;
 
         padding-bottom: 7px;
+        border-bottom: 1px solid #9d9d9d;
 
         .idcircle {
           width: 30px;
