+++ src/components/Find/FindList.jsx
@@ -4,7 +4,7 @@ export default function FindList({ id, nickname }) {
   return (
     <div>
       <li key={id}>
-        <span className="idcircle"></span>
+        <span className="idcircle" />
         <div className="idNames">
           <span className="idName">{id}</span>
           <span className="idNickName">{nickname}</span>
