+++ src/pages/galim/Login/Login.scss
@@ -18,7 +18,7 @@
 
 .login-wrap {
   flex-direction: column;
-  width: 100vw;
+  width: 1100px;
   height: 100vh;
   border: 1px solid #e6e6e6;
   @include flexCenter;
